+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
   - NuvlaBox map view now show all NuvlaBoxes, including NBs with inferred location
+  - Fix refresh of cluster info on default NuvlaBox view
+
 
 ## [2.22.0] - 2021-09-14